Add support for RAILS_RELATIVE_ROOT
If the RAILS_RELATIVE_ROOT is set, Web Console is missing its end-point,
and so it returns a 404 error.

So if the relative root is set, we concatenate it with the default mount point.
Or if the "config.web_console.mount_point" is set as configuration,
we force to use it as mount point of Web Console.

// An example case

```ruby
map '/relative/root' do
  run MyApp
end
```

```shell
$ RAILS_RELATIVE_URL_ROOT=/relative/root bundle exec rails s
```